# app/Http/Requests/UpdateItemRequest.php
<?php

namespace App\Http\Requests;

use Illuminate\Foundation\Http\FormRequest;

/**
 * @OA\Schema(
 *     title="UpdateItemRequest",
 *     description="Update Item Request Schema",
 *     type="object",
 *     required={"name", "image", "price"}
 * )
 */
class UpdateItemRequest extends FormRequest
{
    /**
     * Determine if the user is authorized to make this request.
     */
    public function authorize(): bool
    {
        return in_array(auth()->user()->role_id, [1, 2]);
    }

    /**
     * Get the validation rules that apply to the request.
     *
     * @return array<string, \Illuminate\Contracts\Validation\ValidationRule|array<mixed>|string>
     */
    public function rules(): array
    {
        return [
            'name'  => 'required|string|max:86',
            'image'  => 'required|string|max:18',
            'image_file'  => 'file|mimes:jpg,png',
            'price' => 'required|integer',
        ];
    }
}