# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation POS API",
        "description": "This API documentation outlines the features and endpoints of a Laravel project designed to master RESTful API development using Sanctum. It provides comprehensive documentation for secure user authentication, token-based sessions, and role-based access control. Developed as a learning tool, this project demonstrates best practices in Laravel API design, including resourceful routing, middleware application, and data validation techniques. It serves as a practical guide for developers aiming to implement modern API solutions with Laravel and Sanctum",
        "contact": {
            "email": "pichoendoo@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/items": {
            "get": {
                "tags": [
                    "Items"
                ],
                "summary": "Get items",
                "description": "Fetch a list of items with optional search query",
                "operationId": "getItems",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Items fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Item"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Fetch successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to fetch items. Please try again later."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Items"
                ],
                "summary": "Create item",
                "description": "Create a new item",
                "operationId": "storeItem",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreItemRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Item created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Item"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Item created successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to create item. Please try again later."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/items/{item}": {
            "get": {
                "tags": [
                    "Items"
                ],
                "summary": "Get item",
                "description": "Fetch a specific item by ID",
                "operationId": "getItem",
                "parameters": [
                    {
                        "name": "item",
                        "in": "path",
                        "description": "Item ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Item fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Item"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Fetch successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to fetch item. Please try again later."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Items"
                ],
                "summary": "Update item",
                "description": "Update an existing item",
                "operationId": "updateItem",
                "parameters": [
                    {
                        "name": "item",
                        "in": "path",
                        "description": "Item ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateItemRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Item updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Item"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Item updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to update item. Please try again later."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Items"
                ],
                "summary": "Delete item",
                "description": "Delete a specific item by ID",
                "operationId": "deleteItem",
                "parameters": [
                    {
                        "name": "item",
                        "in": "path",
                        "description": "Item ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Item deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to delete item. Please try again later."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreItemRequest": {
                "title": "StoreItemRequest",
                "description": "Store Item Request",
                "required": [
                    "name",
                    "image",
                    "price",
                    "category_id",
                    "cogs",
                    "qty"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Product Name"
                    },
                    "image": {
                        "type": "string",
                        "example": "image-url.jpg"
                    },
                    "image_file": {
                        "type": "file",
                        "format": "binary"
                    },
                    "price": {
                        "type": "integer",
                        "example": 100
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "cogs": {
                        "type": "string",
                        "example": "Cost of goods sold"
                    },
                    "qty": {
                        "type": "integer",
                        "example": 10
                    }
                },
                "type": "object"
            },
...
            "UpdateItemRequest": {
                "title": "UpdateItemRequest",
                "description": "Update Item Request Schema",
                "required": [
                    "name",
                    "image",
                    "price"
                ],
                "type": "object"
            },
...
            "Item": {
                "title": "Item",
                "description": "Item model",
                "properties": {
                    "id": {
                        "description": "Item ID",
                        "type": "integer",
                        "example": "1"
                    },
                    "uuid": {
                        "description": "UUID of the item",
                        "type": "string",
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    "name": {
                        "description": "Name of the item",
                        "type": "string",
                        "example": "Sample Item"
                    },
                    "image": {
                        "description": "URL to the image of the item",
                        "type": "string",
                        "example": "http://example.com/item.jpg"
                    },
                    "category_id": {
                        "description": "Category ID of the item",
                        "type": "integer",
                        "example": "1"
                    },
                    "price": {
                        "description": "Price of the item",
                        "type": "number",
                        "format": "decimal",
                        "example": "19.99"
                    },
                    "created_at": {
                        "description": "Date and time when the item was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-07-03 12:00:00"
                    },
                    "updated_at": {
                        "description": "Date and time when the item was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-07-03 12:30:00"
                    },
                    "deleted_at": {
                        "description": "Date and time when the item was soft deleted",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-07-03 12:45:00"
                    },
                    "stock_count": {
                        "description": "Computed attribute: Total stock count of the item",
                        "type": "integer",
                        "example": "100"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}